pi/requirements.txt
pyserial
flask
paho-mqtt